move code to wcsmap
